Unqueue elements with minimum and maximum priority
Previously, you had specify the priority of the
bucket from which to unqueue elements. This is no
longer necessary as BucketQueue is now able to
find the bucket with minimum or maximum priority.

The SimpleIndex implements the optimised approach
described on the Wikipedia page for bucket queues.
This implements the Index trait, which means more
sophisticated approaches could be implemented,
for example, using a min-max heap.

The SimpleIndex works by tracking the minimum and
maximum priorities seen so far. When the last
element is removed from these buckets, a
sequential scan through the vector in both
directions determines the new minimum and maximum.