Allow BucketQueue to be nested inside another one
BucketQueue now implements the Bucket trait which
means it can be used as a type of Bucket inside
another BucketQueue.

By implementing this trait, BucketQueue#new, #len
and #is_empty became ambiguous. To disambiguate
these functions, their names have been suffixed
with the trait name.

The interface for using nested BucketQueues is
clunky and also error-prone, because there’s
nothing ensuring the #bucket_for_adding and
#bucket_for_removing functions add or remove
exactly one item. This is important because it
ensures the Index remains consistent.