Simplify access to nested BucketQueues
When BucketQueue is nested, it becomes difficult
to access its inner buckets because you need to
call the ‘bucket_for_adding’ and ‘_removing’
functions and these are implementation details
that the programmer shouldn’t need to worry about.

This commit introduces a new Deferred trait and a
DeferredBucket struct that simplify things. You
can now call ‘bucket’ on a BucketQueue to build a
DeferredBucket that manages future access to the
bucket, meaning you can do this:

```
bucket_queue.bucket(0).bucket(1).push(“foo”, 2)
```

This is easier to deal with and chains nicely.
However, there are two downsides to this approach:

1. The DeferredBucket is a struct and building it
   could add a small amount of overhead.

2. Any new queue or bucket types must implement
   Deferred if they want to make use of it.

There’s also considerable complexity in getting
these traits to work in the first place. The
constraints are rather complex and hard to read.