Add support for clearing and pruning BucketQueue
The Rust convention is that #clear removes
everything from the data structure, which is
followed here. I’ve also implemented prune, which
removes everything from one of the buckets in the
BucketQueue.

This poses a problem when pruning items from a
nested BucketQueue because the top-level
BucketQueue needs to update its index to reflect
the new length of items in the queue.

Additionally, if the nested BucketQueue is now
empty after pruning, the top-level queue might
need to update its min and max indexes.

These problems are solved by having the
DeferredBucket call #items_pruned on the parent
queue to inform it how many items were pruned.
The index is then updated with the new length and
values for min and max.